Refractoring and some ESLint warnings resolved
