+++ src/global.d.ts
@@ -10,5 +10,5 @@ declare module "*FeelingsAPI";
 declare module "*GoalsAPI";
 
 export type feelingListType = {
-  Date: IFeelingItem[];
+  [Date as string]: IFeelingItem[];
 };
