+++ src/pages/FeedbackPage/FeedbackPage.tsx
@@ -1,10 +1,9 @@
-/* eslint-disable no-alert */
 import React, { useState } from "react";
-import { Container, Button } from "react-bootstrap";
 import { useTranslation } from "react-i18next";
+import { Container, Button } from "react-bootstrap";
 import { useRecoilValue, useSetRecoilState } from "recoil";
 
-import { darkModeState, displayLoader } from "@store";
+import { darkModeState, displayLoader, displayToast } from "@store";
 import { submitFeedback } from "@src/api/FeedbackAPI";
 import { LandingHeader } from "@components/HeaderDashboard/LandingHeader";
 
@@ -16,6 +15,7 @@ export const FeedbackPage = () => {
   const [userFeedback, setUserFeedback] = useState("");
   const darkModeStatus = useRecoilValue(darkModeState);
   const setLoading = useSetRecoilState(displayLoader);
+  const setDisplayToast = useSetRecoilState(displayToast);
 
   async function submitToAPI(feedback: string) {
     const updatedFeedback = `Rating : ${userRating}\n${feedback}`;
@@ -23,18 +23,16 @@ export const FeedbackPage = () => {
     const res = await submitFeedback(updatedFeedback);
     setLoading(false);
     if (res.status === "success") {
-      alert(res.message);
       setUserFeedback("");
       setUserRating(0);
-    } else {
-      alert(res.message);
     }
+    setDisplayToast({ open: true, message: res.message, extra: "" });
   }
   const { t } = useTranslation();
 
   return (
     <div id="feedback-container">
-      <LandingHeader avatar={"back"} />
+      <LandingHeader avatar="back" />
       <Container className="slide" fluid>
         {userRating === 0 ? <h1>hello</h1> : null}
         <div style={{ color: `${darkModeStatus ? "white" : "black"}` }}>
