+++ src/pages/InvitePage/InvitePage.tsx
@@ -1,18 +1,26 @@
-import React, { useEffect, useState } from "react";
-import { useNavigate } from "react-router";
 import { useRecoilValue } from "recoil";
+import { useNavigate } from "react-router";
+import React, { useEffect, useState } from "react";
 
+import { darkModeState } from "@src/store";
 import { addContact } from "@src/api/ContactsAPI";
 import { queryStyle } from "@src/constants/booleanScreen";
-import { darkModeState } from "@src/store";
-import { LandingHeader } from "@components/HeaderDashboard/LandingHeader";
 import { acceptRelationship } from "@src/services/contact.service";
+import { LandingHeader } from "@components/HeaderDashboard/LandingHeader";
 
 const InvitePage = () => {
   const navigate = useNavigate();
   const darkModeStatus = useRecoilValue(darkModeState);
   const [newContactName, setNewContactName] = useState("");
 
+  const handleSubmit = async () => {
+    const res = await acceptRelationship();
+    if (res.success) {
+      await addContact(newContactName, res.response?.relId, true);
+      setNewContactName("");
+    }
+    navigate("/");
+  };
   useEffect(() => {
     const checkin = localStorage.getItem("checkedIn");
     if (!checkin) {
@@ -31,19 +39,12 @@ const InvitePage = () => {
         The person that sent you this, wants to connect with you here.
         <br />Add them to your contacts.
       </p>
-      <input onChange={(e) => setNewContactName(e.target.value)} className={`default-input${darkModeStatus ? "-dark" : ""}`} placeholder="Name" />
+      <input onSubmit={handleSubmit} onChange={(e) => setNewContactName(e.target.value)} className={`default-input${darkModeStatus ? "-dark" : ""}`} placeholder="Name" />
       <button
         type="button"
         className={`default-btn${darkModeStatus ? "-dark" : ""}`}
         style={{ alignSelf: "right" }}
-        onClick={async () => {
-          const res = await acceptRelationship();
-          if (res.success) {
-            await addContact(newContactName, res.response?.relId, true);
-            setNewContactName("");
-          }
-          navigate("/");
-        }}
+        onClick={handleSubmit}
       > Add Contact
       </button>
     </div>
