+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { Container, Row, Button, Col } from "react-bootstrap";
 import { useRecoilValue } from "recoil";
 import { useNavigate } from "react-router-dom";
+import { Container, Row, Button, Col } from "react-bootstrap";
 
 import { darkModeState } from "@store";
